Initial (broken) refactoring.
