--- src/hydra/UI/__init__.py
@@ -1,3 +1,0 @@
-"""User interface submodule. Defines all UI classes, addon settings and preferences."""
-
-#For AutoAPI documentation.
