+++ src/hydra/addon/__init__.py
@@ -1,0 +1,9 @@
+"""User interface submodule. Defines all UI classes, addon settings and preferences."""
+
+from Hydra.addon import preferences, properties, ops_heightmap, ops_object, ops_image, ui_image, ui_object, ui_common
+
+def get_exports():
+    ret = []
+    for module in (preferences, properties, ops_heightmap, ops_object, ops_image, ui_image, ui_object, ui_common):
+        ret += module.get_exports()
+    return ret
