+++ src/hydra/utils/nav.py
@@ -5,7 +5,7 @@ from Hydra import common
 from mathutils import Euler
 import math
 
-def getSplitDir(target)->str:
+def get_split_direction(target)->str:
 	"""Creates a split direction based on the window size and global preferences.
 
 	:param target: Window to be split.
@@ -19,7 +19,7 @@ def getSplitDir(target)->str:
 	else:
 		return "VERTICAL" if target.height < target.width else "HORIZONTAL"
 
-def getOrMakeArea(type: str, uiType: str = "")->bpy.types.Area:
+def get_or_make_area(type: str, uiType: str = "")->bpy.types.Area:
 	"""Gets or creates a window of the specified type.
 	
 	:param type: Window type.
@@ -38,7 +38,7 @@ def getOrMakeArea(type: str, uiType: str = "")->bpy.types.Area:
 					break
 		if not other:	#splits active
 			area = next(i for i in bpy.context.screen.areas if i.spaces[0] == active)
-			bpy.ops.screen.area_split(direction=getSplitDir(area))
+			bpy.ops.screen.area_split(direction=get_split_direction(area))
 			other = bpy.context.screen.areas[-1]
 		return other
 
@@ -57,7 +57,7 @@ def getOrMakeArea(type: str, uiType: str = "")->bpy.types.Area:
 		target = bpy.context.screen.areas[-1]
 	
 	with bpy.context.temp_override(area=target): #select target area for split
-		bpy.ops.screen.area_split(direction=getSplitDir(target))
+		bpy.ops.screen.area_split(direction=get_split_direction(target))
 
 	target = bpy.context.screen.areas[-1]	#new area is last
 	target.type = type
@@ -66,16 +66,16 @@ def getOrMakeArea(type: str, uiType: str = "")->bpy.types.Area:
 	return target
 
 
-def gotoImage(img: bpy.types.Image):
+def goto_image(img: bpy.types.Image):
 	"""Navigates Blender to the specified image.
 	
 	:param img: Image to navigate to.
 	:type img: :class:`bpy.types.Image`"""
-	imgEditor = getOrMakeArea("IMAGE_EDITOR")
+	imgEditor = get_or_make_area("IMAGE_EDITOR")
 	space = imgEditor.spaces[0]
 	space.image = img
 
-def gotoShader(obj: bpy.types.Object):
+def goto_shader(obj: bpy.types.Object):
 	"""Navigates Blender to the material of the specified object.
 	
 	:param obj: Object to navigate to.
@@ -84,36 +84,36 @@ def gotoShader(obj: bpy.types.Object):
 	if (len(mats) == 0):
 		return
 	
-	nodeEditor = getOrMakeArea("NODE_EDITOR", "ShaderNodeTree")
+	nodeEditor = get_or_make_area("NODE_EDITOR", "ShaderNodeTree")
 	space = nodeEditor.spaces[0]
 	space.shader_type = "OBJECT"
 
-def gotoObject(obj: bpy.types.Object):
+def goto_object(obj: bpy.types.Object):
 	"""Navigates Blender to the specified object.
 	
 	:param obj: Object to navigate to.
 	:type obj: :class:`bpy.types.Object`"""
-	space = getOrMakeArea("VIEW_3D").spaces[0]
+	space = get_or_make_area("VIEW_3D").spaces[0]
 	space.region_3d.view_rotation = Euler((3/math.pi,0,math.pi/4), "XYZ").to_quaternion()
 	space.region_3d.view_location = obj.location
 
-def gotoModifier():
+def goto_modifier():
 	"""Navigates to the Modifiers tab."""
-	space = getOrMakeArea("PROPERTIES").spaces[0]
+	space = get_or_make_area("PROPERTIES").spaces[0]
 	space.context = "MODIFIER"
 
-def gotoShape():
+def goto_shape():
 	"""Navigates to the Mesh data tab."""
-	space = getOrMakeArea("PROPERTIES").spaces[0]
+	space = get_or_make_area("PROPERTIES").spaces[0]
 	space.context = "DATA"
 
-def gotoGeometry(obj: bpy.types.Object):
+def goto_geometry(obj: bpy.types.Object):
 	"""Navigates to the Geometry Nodes tab."""
 	nodes = [m for m in obj.modifiers if m.type == "NODES"]
 	if len(nodes) == 0 or nodes[-1].node_group is None:
 		return
 	
-	nodeEditor = getOrMakeArea("NODE_EDITOR", "GeometryNodeTree")
+	nodeEditor = get_or_make_area("NODE_EDITOR", "GeometryNodeTree")
 	space = nodeEditor.spaces[0]
 	space.geometry_nodes_type = "MODIFIER"
 	space.node_tree = nodes[-1].node_group
